# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GPU Upsampler Control",
    "description": "\n## GPU Audio Upsampler Web API\n\nControl interface for the GPU-accelerated audio upsampler daemon.\n\n### Features\n- **Daemon Control**: Start/stop/restart the audio processing daemon\n- **EQ Management**: Import and manage EQ profiles (AutoEq/Equalizer APO format)\n- **OPRA Integration**: Access the OPRA headphone EQ database\n- **Real-time Stats**: WebSocket endpoint for live statistics\n\n### Authentication\nNo authentication required (local network only).\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/system/logs": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get System Logs",
        "description": "Retrieve system logs from the daemon log file.\n\nArgs:\n    lines: Number of lines to return (1-1000)\n    offset: Number of lines to skip from the beginning\n    level: Optional log level filter\n\nReturns:\n    SystemLogsResponse containing logs and metadata\n\nRaises:\n    HTTPException: If log file is too large or cannot be read",
        "operationId": "get_system_logs_api_system_logs_get",
        "parameters": [
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of lines to return",
              "default": 100,
              "title": "Lines"
            },
            "description": "Number of lines to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of lines to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of lines to skip"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by log level (debug, info, warning, error)",
              "title": "Level"
            },
            "description": "Filter by log level (debug, info, warning, error)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemLogsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/eq": {
      "get": {
        "summary": "Eq Page",
        "description": "Serve the EQ Settings page.",
        "operationId": "eq_page_eq_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crossfeed": {
      "get": {
        "summary": "Crossfeed Page",
        "description": "Serve the Crossfeed page.",
        "operationId": "crossfeed_page_crossfeed_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rtp": {
      "get": {
        "summary": "Rtp Management Page",
        "description": "Serve the RTP Management page.",
        "operationId": "rtp_management_page_rtp_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/system": {
      "get": {
        "summary": "System Page",
        "description": "Serve the System page.",
        "operationId": "system_page_system_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SystemLogsResponse": {
        "properties": {
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs",
            "description": "Array of log lines"
          },
          "total_lines": {
            "type": "integer",
            "title": "Total Lines",
            "description": "Total number of lines in the log file"
          },
          "file_path": {
            "type": "string",
            "title": "File Path",
            "description": "Path to the log file"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "Size of the log file in bytes"
          }
        },
        "type": "object",
        "required": [
          "logs",
          "total_lines",
          "file_path",
          "file_size"
        ],
        "title": "SystemLogsResponse",
        "description": "System logs response model."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}